+++ CHANGELOG.md
@@ -15,6 +15,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.108.0]
+
+### Added
+- DataPicker: Added settings for displaying the input date and time
+- DataPicker: Added the ability to format the result
+- Added code editor for primitivesLibrary field
+- DynamicFormDash
+- Added the ability to assign an icon via metadata
+- Added ability to correct color via metadata
+- ConstructorSchemes: Added loading svg icons from server
+- ConstructorSchemes: Added data source support
+- Added possibility save multiLine in png
+- Text Area: added read only mode
+
+### Changed
+- Renaming the single value component
+- Removed the board for textarea
+
+### Fixed
+- fixed update of tokens on the map
+- fixed position icons\buttons on dash panel
+- fixed flickering map
+
 ## [1.107.0]
 
 ### Added